Server/requirements.txt
asgiref==3.2.3
astroid==2.3.3
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.1
Django==3.0.3
django-extensions==2.2.6
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.11.0
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
ipython==7.12.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.1
parso==0.6.1
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
python-decouple==3.3
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.7
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
asgiref==3.2.3
astroid==2.3.3
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.1
decouple==0.0.7
Django==3.0.3
django-extensions==2.2.6
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.11.0
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
ipython==7.12.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.1
parso==0.6.1
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.7
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
